rename: fix typo
